LayoutTests/fast/forms/radio/radio_checked_dynamic-expected.txt

This will test that radio buttons can be moved in and out of forms while maintaining only one checked button per group, per form.
Test 1: Transfer radio buttons from <body> to <form>:

Test 1 Passed

Test 2: Transfer radio buttons from <form> to <body>:
Test 2 Passed

Test 3: Transfer radio buttons from <form> to <body> with a dormant state in the middle:
Test 3 Passed

Test 4: Transfer from <form> to <form>

Test 4 Passed
